refactor(workitems): extract query into nx_lib/workitems
Move the workitems list-query engine (the 395-line _get_workitems_data
body: the two doc-field pre-resolution legs, WorkitemFilter construction,
fetch_merged_page) and the value-first docfield-suggestions core into
nx_lib/workitems/query.py as get_workitems_data/docfield_suggestions_any_field
-- pure Python, no Flask/session/request/current_app (D4). ms02_pid_specs/
ms02_prepared_docs_processes (already session-free, just mapping_config
reads) move alongside it. Moved LAST within the package, after
fields/sensitivity/media, since it is built on top of all three.

Every session/permission read the interactive list path needs is still
gathered by _session_scope() in views/workitems.py (unchanged, Flask-aware);
every DB engine, resolver callable, cache, and the app logger is now passed
into the pure core explicitly instead of being read as a Flask-context
global. get_workitems_data itself never writes to session -- it returns the
resolved process-selection string for the caller to persist, and never
stamps MS02 pid/in_register rows (that stays _stamp_in_register's job, in
views/workitems.py, since it is itself session/current_app-coupled).

views/workitems.py keeps every moved name's original signature and
call-site behaviour as thin wrappers that read this module's OWN globals
(engines, resolvers, get_valid_search_columns(), current_app.logger) and
forward them explicitly -- this is what keeps the large existing
monkeypatch-based test suite (which patches names like engine_statistics_db,
resolve_ms02_docfield_ids, get_valid_search_columns, has_permission as
module attributes on nx_lib.views.workitems) passing unchanged. The one
exception is the _docfield_op spy test (test_workitems_routes.py), updated
in the fields.py commit to patch nx_lib.workitems.fields directly, since the
pair-fold loop now calls it module-qualified from there.

Added unit tests in tests/unit/test_workitems_query.py exercising
get_workitems_data/docfield_suggestions_any_field/ms02_pid_specs/
ms02_prepared_docs_processes directly with fakes for the DB engines, cache,
fetch_merged_page, and mapping_config -- no database, no Flask app context.
